moves ir into frontend crate
