+++ crates/backend/src/wasm_builder.rs
@@ -2,8 +2,8 @@ use std::fmt::Write;
 use std::iter;
 use std::{collections::HashMap, mem};
 
-use crate::ir::{FuncTy, Id, Import, Name, Struct, Substitution, Ty, Variant};
-use crate::names::NameSupply;
+use frontend::ir::{FuncTy, Id, Import, Name, Struct, Substitution, Ty, Variant};
+use frontend::names::NameSupply;
 use wasm_encoder::{
     self, ArrayType, CodeSection, CompositeType, ConstExpr, ElementSection, Elements, EntityType,
     ExportKind, ExportSection, FieldType, FuncType, Function, FunctionSection, GlobalSection,
