+++ crates/cli/Cargo.toml
@@ -10,4 +10,5 @@ path = "src/main.rs"
 [dependencies]
 clap = { version = "4.4.12", features = ["derive"] }
 frontend = { path = "../frontend" }
+backend = { path = "../backend" }
 language-server = { path = "../language-server" }
