+++ crates/cli/src/main.rs
@@ -1,7 +1,8 @@
 use std::{error::Error, fs, path::PathBuf, process};
 
 use clap::{Parser, Subcommand};
-use frontend::{compile_program, render_errors};
+use frontend::render_errors;
+use backend::compile_program;
 use language_server::start_language_server;
 
 /// The nemo language
