+++ crates/frontend/Cargo.toml
@@ -12,7 +12,6 @@ num-derive = "0.4.2"
 num-traits = "0.2.18"
 rowan = "0.15.15"
 lazy_static = "1.4.0"
-backend = { path = "../backend" }
 text-size = "1.1.1"
 line-index = "0.1.1"
