+++ crates/frontend/src/types.rs
@@ -8,8 +8,8 @@ use self::{
     errors::TyError,
 };
 use crate::syntax::{nodes::Root, token_ptr::SyntaxTokenPtr, SyntaxNodePtr};
-pub use backend::ir::{FuncTy, NameMap, Program, Ty};
-use backend::ir::{Name, NameSupply};
+pub use crate::ir::{FuncTy, NameMap, Program, Ty};
+use crate::ir::{Name, NameSupply};
 use std::collections::HashMap;
 
 #[derive(Debug)]
